+++ tests/fixtures/mini-fastify/lib/content-type-parser.js
@@ -1,0 +1,7 @@
+'use strict'
+
+function addContentTypeParser (fastify, contentType, parser) {
+  fastify.addContentTypeParser(contentType, parser)
+}
+
+module.exports = { addContentTypeParser }